Update to ChoiceboxGroup
