--- src/components/input-elements/RadioGroup/index.ts
@@ -1,4 +1,0 @@
-export { default as RadioGroup } from "./RadioGroup";
-export type { RadioGroupProps } from "./RadioGroup";
-export { default as RadioGroupOption } from "./RadioGroupOption";
-export type { RadioGroupOptionProps } from "./RadioGroupOption";
